Rename to local_hash_sign.py to local_message_signer.py and add middlware utils
